vkconfig: move HasPreset to a more consistent location
