+++ vkconfig_core/layer_preset.cpp
@@ -36,3 +36,16 @@ const LayerPreset* GetPreset(const std::vector<LayerPreset>& presets, const char
 
     return nullptr;
 }
+
+bool HasPreset(const std::vector<LayerSettingData>& layer_settings, const std::vector<LayerSettingData>& preset_settings) {
+    if (preset_settings.empty()) return false;
+
+    for (std::size_t preset_index = 0, preset_count = preset_settings.size(); preset_index < preset_count; ++preset_index) {
+        const LayerSettingData* layer_setting = FindByKey(layer_settings, preset_settings[preset_index].key.c_str());
+        if (layer_setting == nullptr) return false;
+
+        if (preset_settings[preset_index].value != layer_setting->value) return false;
+    }
+
+    return true;
+}
