+++ vkconfig_core/layer_preset.h
@@ -35,3 +35,7 @@ struct LayerPreset {
 };
 
 const LayerPreset* GetPreset(const std::vector<LayerPreset>& presets, const char* preset_label);
+
+// Check whether "layer_settings" has all the settings set in "preset_settings"
+// "layer_settings" may have more settings then "preset_settings" and return true
+bool HasPreset(const std::vector<LayerSettingData>& layer_settings, const std::vector<LayerSettingData>& preset_settings);
